build: Create a spirv_tools_utils.h file
