+++ layers/utils/shader_utils.cpp
@@ -19,16 +19,10 @@
 
 #include "shader_utils.h"
 
-#include "generated/device_features.h"
-#include "generated/vk_api_version.h"
-#include "generated/vk_extension_helper.h"
-#include "utils/hash_util.h"
-
 #include "generated/spirv_tools_commit_id.h"
 
 #include <cstring>
 #include <fstream>
-#include <sstream>
 
 void ValidationCache::GetUUID(uint8_t *uuid) {
     const char *sha1_str = SPIRV_TOOLS_COMMIT_ID;
@@ -114,113 +108,6 @@ void ValidationCache::Merge(ValidationCache const *other) {
     auto guard = WriteLock();
     good_shader_hashes_.reserve(good_shader_hashes_.size() + other->good_shader_hashes_.size());
     for (auto h : other->good_shader_hashes_) good_shader_hashes_.insert(h);
-}
-
-spv_target_env PickSpirvEnv(const APIVersion &api_version, bool spirv_1_4) {
-    if (api_version >= VK_API_VERSION_1_4) {
-        return SPV_ENV_VULKAN_1_4;
-    } else if (api_version >= VK_API_VERSION_1_3) {
-        return SPV_ENV_VULKAN_1_3;
-    } else if (api_version >= VK_API_VERSION_1_2) {
-        return SPV_ENV_VULKAN_1_2;
-    } else if (api_version >= VK_API_VERSION_1_1) {
-        if (spirv_1_4) {
-            return SPV_ENV_VULKAN_1_1_SPIRV_1_4;
-        } else {
-            return SPV_ENV_VULKAN_1_1;
-        }
-    }
-    return SPV_ENV_VULKAN_1_0;
-}
-
-// Some Vulkan extensions/features are just all done in spirv-val behind optional settings
-void AdjustValidatorOptions(const DeviceExtensions &device_extensions, const DeviceFeatures &enabled_features,
-                            spv_target_env spirv_environment, spvtools::ValidatorOptions &out_options, uint32_t *out_hash,
-                            std::string &out_command) {
-    struct Settings {
-        bool relax_block_layout;
-        bool uniform_buffer_standard_layout;
-        bool scalar_block_layout;
-        bool workgroup_scalar_block_layout;
-        bool allow_local_size_id;
-        bool allow_offset_texture_operand;
-        bool allow_vulkan_32_bit_bitwise;
-    } settings;
-
-    // VK_KHR_relaxed_block_layout never had a feature bit so just enabling the extension allows relaxed layout
-    // Was promotoed in Vulkan 1.1 so anyone using Vulkan 1.1 also gets this for free
-    settings.relax_block_layout = IsExtEnabled(device_extensions.vk_khr_relaxed_block_layout);
-    // The rest of the settings are controlled from a feature bit, which are set correctly in the state tracking. Regardless of
-    // Vulkan version used, the feature bit is needed (also described in the spec).
-    settings.uniform_buffer_standard_layout = enabled_features.uniformBufferStandardLayout == VK_TRUE;
-    settings.scalar_block_layout = enabled_features.scalarBlockLayout == VK_TRUE;
-    settings.workgroup_scalar_block_layout = enabled_features.workgroupMemoryExplicitLayoutScalarBlockLayout == VK_TRUE;
-    settings.allow_local_size_id = enabled_features.maintenance4 == VK_TRUE;
-    settings.allow_offset_texture_operand = enabled_features.maintenance8 == VK_TRUE;
-    settings.allow_vulkan_32_bit_bitwise = enabled_features.maintenance9 == VK_TRUE;
-
-    std::stringstream ss;
-    ss << "spirv-val <input.spv>";
-
-    if (settings.relax_block_layout) {
-        ss << " --relax-block-layout";
-        out_options.SetRelaxBlockLayout(true);
-    }
-    if (settings.uniform_buffer_standard_layout) {
-        ss << " --uniform-buffer-standard-layout";
-        out_options.SetUniformBufferStandardLayout(true);
-    }
-    if (settings.scalar_block_layout) {
-        ss << " --scalar-block-layout";
-        out_options.SetScalarBlockLayout(true);
-    }
-    if (settings.workgroup_scalar_block_layout) {
-        ss << " --workgroup-scalar-block-layout";
-        out_options.SetWorkgroupScalarBlockLayout(true);
-    }
-    if (settings.allow_local_size_id) {
-        ss << " --allow-localsizeid";
-        out_options.SetAllowLocalSizeId(true);
-    }
-    if (settings.allow_offset_texture_operand) {
-        ss << " --allow-offset-texture-operand";
-        out_options.SetAllowOffsetTextureOperand(true);
-    }
-    if (settings.allow_vulkan_32_bit_bitwise) {
-        ss << " --allow-vulkan-32-bit-bitwise";
-        out_options.SetAllowVulkan32BitBitwise(true);
-    }
-
-    switch (spirv_environment) {
-        case SPV_ENV_VULKAN_1_4:
-            ss << " --target-env vulkan1.4";
-            break;
-        case SPV_ENV_VULKAN_1_3:
-            ss << " --target-env vulkan1.3";
-            break;
-        case SPV_ENV_VULKAN_1_2:
-            ss << " --target-env vulkan1.2";
-            break;
-        case SPV_ENV_VULKAN_1_1:
-            ss << " --target-env vulkan1.1";
-            break;
-        case SPV_ENV_VULKAN_1_0:
-            ss << " --target-env vulkan1.0";
-            break;
-        default:
-            break;
-    }
-
-    // Faster validation without friendly names.
-    out_options.SetFriendlyNames(false);
-
-    // The spv_validator_options_t in libspirv.h is hidden so we can't just hash that struct, so instead need to create our own.
-    if (out_hash) {
-        *out_hash = hash_util::Hash32(&settings, sizeof(Settings));
-    }
-
-    ss << "\n";
-    out_command = ss.str();
 }
 
 // This is used to help dump SPIR-V while debugging intermediate phases of any altercations to the SPIR-V
